add shade plugin to build fat jar

+++ classifier/src/main/java/in/nimbo/entity/Data.java
@@ -1,22 +1,22 @@
 package in.nimbo.entity;
 
 public class Data {
-    private long label;
+    private double label;
     private String content;
 
     public Data() {}
 
     public Data(long label, String content) {
-        this.label = label;
+        this.label = (double) label;
         this.content = content;
     }
 
-    public long getLabel() {
+    public double getLabel() {
         return label;
     }
 
     public void setLabel(long label) {
-        this.label = label;
+        this.label = (double)label;
     }
 
     public String getContent() {
